add fixture admin category

+++ src/Service/PostService.php
@@ -2,57 +2,65 @@
 
 namespace App\Service;
 
+use App\Dto\NewCategoryDto;
 use App\Dto\NewPostDto;
+use App\Entity\Category;
 use App\Entity\Post;
+use App\Repository\PostRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 class PostService
 {
     public function __construct(
         public EntityManagerInterface $em,
         public Security $security,
+        public PostRepository $postRepository,
+        public AuthorizationCheckerInterface $authorizationChecker,
     ) {}
 
     public function showPost()
     {
-        $posts =  $this->em->getRepository('App:Post')->findAll();
-
-        if (!$posts) {
-            throw new NotFoundHttpException("not exist");
-        }
-
-        return $posts;
+        return $this->postRepository->findAll();
     }
 
     public function showPostById(int $id)
     {
-         $post = $this->em->getRepository('App:Post')->find($id);
-
-        if (!$post) {
-            throw new NotFoundHttpException("not exist");
-        }
-
-        return $post;
+        return $this->postRepository->find($id);
     }
 
-    public function newPost(NewPostDto $newPostDto)
+    public function newPost(NewPostDto $newPostDto, NewCategoryDto $newCategoryDto)
     {
         $post = new Post();
+        $category = new Category();
+
+        if(!$this->authorizationChecker->isGranted('new', $post)) {
+            throw new AccessDeniedException('Posts can be create by user or admin');
+        }
+
+        $category->setName($newCategoryDto->getName());
+
         $post->setTitle($newPostDto->getTitle())
-            ->setContent($newPostDto->getContent())
+            ->setText($newPostDto->getText())
             ->setAuthor($this->security->getUser())
+            ->addCategory($category)
         ;
 
+        $this->em->persist($category);
         $this->em->persist($post);
         $this->em->flush();
     }
 
     public function editPost(NewPostDto $newPostDto, Post $post)
     {
+        if(!$this->authorizationChecker->isGranted('edit', $post)) {
+            throw new AccessDeniedException('Posts can be edit by author or admin');
+        }
+
         $post->setTitle($newPostDto->getTitle())
-            ->setContent($newPostDto->getContent())
+            ->setText($newPostDto->getText())
             ->setAuthor($this->security->getUser())
         ;
 
@@ -61,6 +69,10 @@ class PostService
 
     public function removePost(Post $post)
     {
+        if(!$this->authorizationChecker->isGranted('delete', $post)) {
+            throw new AccessDeniedException('Posts can be remove by author or admin');
+        }
+
         $this->em->remove($post);
         $this->em->flush();
     }
